+++ src/timecodedisplay.cpp
@@ -65,7 +65,7 @@ TimecodeDisplay::TimecodeDisplay(const Timecode &t, QWidget *parent)
     setTimeCodeFormat(KdenliveSettings::frametimecode(), true);
     setValue(m_minimum);
     setMinimumWidth(fm.width(QStringLiteral("88:88:88:88")) + contentsMargins().right() + contentsMargins().left() + frameSize().width() -
-                    lineEdit()->contentsRect().width() + QStyle::PM_SpinBoxFrameWidth + 6);
+                    lineEdit()->contentsRect().width() + (int)QStyle::PM_SpinBoxFrameWidth + 6);
 
     setSizePolicy(QSizePolicy::Fixed, QSizePolicy::Maximum);
     setAccelerated(true);
